add regression test for #280
This test dies with an overflow. With CHALK_DEBUG=1 this death is
quick, without it, it is slow and painful. The problem is that some of
the types in the environment are "truncated" and replaced with
inference variables, but those types are unnormalized
projections. When we then go to unify the environment found in the
answer, we end up needing to unify inference variables with
unnormalized projections -- but the current code creates subgoals
instead, which leads to an infinite recursion.