have unification produce *goals* rather than constraints
Solving those goals leads to the same constraints being generated as
before, though sometimes in a slightly different order.

One interesting diff is `equality_binder1` -- here, the universe of a
variable also changed from U1 to U0. But if you look at the structure
of the answer, there is no difference -- this variable is only
affected by things in U2. This falls out from universe
canonicalization -- though it could maybe be considered a bug.  I
believe it happens because we don't see a need to persist U1 when
universe canonicalizing, just U0 and U2, and hence U1 gets mapped to
U0. I suppose that this could be a problem in that the variable does
appear as part of the result substitution, and losing detail on the
proper universe might lead to incorrect errors later.