extract placeholders from a type-name into their own variant
Motivation #1: We want type-names to be opaquely defined by the type
family. The fact that we had to fold `ApplicationTy` into `Ty` should
have been a sign that something should change.

Motivation #2: Although this commit does get that far, we want to
treat associated type placeholders akin to ordinary ones.